% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atcol.R
\name{atcol}
\alias{atcol}
\title{atcol()}
\usage{
atcol(x, j, n)
}
\arguments{
\item{x}{A sudoku grid}

\item{j}{An index of a column from the grid}

\item{n}{a value to check its existance in the column j}
}
\value{
TRUE if the checked value is on the column or FALSE if the checked value is not on the column
}
\description{
This function checks if a value already exists in a column from a sudoku grid
}
\examples{
atcol(x=grid_gen(63),1,8)
atcol(x=grid_gen(49),7,6)
}
