% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_solver.R
\name{bt_solver}
\alias{bt_solver}
\title{bt_solver()}
\usage{
bt_solver(x)
}
\arguments{
\item{x}{A sudoku grid}
}
\value{
A list of two elements in the first one there is the grid x solved as a matrix of 9 by 9, and the second one contains the number of backtracking does R do to solving it.
}
\description{
This function is a recurcive function that solves a sudoku grid using the backtracking algorithme
}
\examples{
bt_solver(x=grid_gen(49))
}
