% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grid.R
\name{check_grid}
\alias{check_grid}
\title{Check_grid()}
\usage{
check_grid(x)
}
\arguments{
\item{x}{A sudoku grid}
}
\value{
True if x is a cpmlete sudoku grid False if x is not
}
\description{
This function checks if a 9 by 9 grid is a complete sudoku grid (each number appear only once in its row,column and box )
}
\examples{
check_grid(x=grid_gen_cplt())
check_grid(x=grid_gen(54))
}
