% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_gen.R
\name{grid_gen}
\alias{grid_gen}
\title{grid_gen()}
\usage{
grid_gen(t)
}
\arguments{
\item{t}{The number of the emty cells}
}
\value{
A sudoku grid with \code{t} empty cells
}
\description{
This function generates a sudoku grid with a given number for the emty cells
}
\examples{
Grid_45 = grid_gen(45)
}
