% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_gen_cplt.R
\name{grid_gen_cplt}
\alias{grid_gen_cplt}
\title{grid_gen_cplt()}
\usage{
grid_gen_cplt()
}
\value{
A complete sudoku grid
}
\description{
This function generates a complete sudoku grid randomly
}
\examples{
Grid_complete = grid_gen_cplt()
}
