% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_gen_lv.R
\name{grid_gen_lv}
\alias{grid_gen_lv}
\title{grid_gen_lv()}
\usage{
grid_gen_lv(lv)
}
\arguments{
\item{lv}{A string argument level for the grid and must be "Easy", "Difficult","Hard" or "Legend"}
}
\value{
A sudoku grid associate to the level in \code{lv}
}
\description{
This function generates a sudoku grid for four levels of playing "Easy", "Difficult","Hard" and "Legend" based
 on the number of backtraking does the finction \code{\link[RSDK]{bt_solver}} did to solve the grid.
}
\examples{
grid_gen_lv("Easy")
grid_gen_lv("Legend")
}
