% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispossible.R
\name{ispossible}
\alias{ispossible}
\title{ispossible()}
\usage{
ispossible(x, i, j, n)
}
\arguments{
\item{x}{A sudoku grid}

\item{i}{The index of the row of the given cell}

\item{j}{The index of the column of the given cell}

\item{n}{The number that we want to check if is possible to put it in the cell of the index (i,j)}
}
\value{
True if it is possible to put \code{n} in the cell (i,j)
}
\description{
This function checks if it is possible to put a given number in a given empty cell
}
\examples{
ispossible(x=grid_gen_cplt(),4,5,6)
ispossible(x=grid_gen_cplt(),4,5,6)
}
