% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbrposs.R
\name{nbrposs}
\alias{nbrposs}
\title{nbrposs()}
\usage{
nbrposs(x, i, j)
}
\arguments{
\item{x}{A sudoku grid}

\item{i}{The index of the row of the given cell}

\item{j}{The index of the column of the given cell}
}
\value{
Number of possibilities for the cell (i,j)
}
\description{
This function returns the number of possibilities for a given empty cell
}
\examples{
nbrposs(x=grid_gen_cplt(),5,7)
nbrposs(x=grid_gen_cplt(),6,9)
}
