% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_vec.R
\name{perm_vec}
\alias{perm_vec}
\title{perm_vec()}
\usage{
perm_vec(x, i)
}
\arguments{
\item{x}{A vector}

\item{i}{The length of the cyclic permutation}
}
\value{
A vector permuted cyclically by \code{x} values
}
\description{
This function permutes a given vector with a cyclic permutaion
}
\examples{
perm_vec(1:6,4)
perm_vec(27:50,15)
}
