% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_grid.R
\name{plt_grid}
\alias{plt_grid}
\title{plt_grid()}
\usage{
plt_grid(X)
}
\arguments{
\item{X}{A sudoku grid}
}
\value{
a plot of the grid
}
\description{
This function plots a given sudoku grid
}
\examples{
plt_grid(X=grid_gen_cplt())
}
