% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_grid_play.R
\name{plt_grid_play}
\alias{plt_grid_play}
\title{plt_grid_play()}
\usage{
plt_grid_play(B, x)
}
\arguments{
\item{B}{Initial grid}

\item{x}{The grid that the user put the numbers on it}
}
\value{
a plot of the grid with the user input with a different color red if the input is on the wrong cell and green if the input is on the right cell
}
\description{
This function gives a reactive plot of the grid for the shiny application
}
