% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poss.R
\name{poss}
\alias{poss}
\title{poss()}
\usage{
poss(x, i, j)
}
\arguments{
\item{x}{A sudoku grid}

\item{i}{The index of the row of the given cell}

\item{j}{The index of the column of the given cell}
}
\value{
Vector of possibilities for the cell (i,j)
}
\description{
This function returns a vector of possibilities for a given empty cell
}
\examples{
poss(x=grid_gen(46),4,7)
poss(x=grid_gen(49),3,9)
}
